+++ TP_VAM_MODI.sln
@@ -3,26 +3,46 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VamLib", "VamLib\VamLib.csproj", "{B131C715-66E0-4DB4-BB2C-0AB6156CE52E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TpSolver", "TpSolver\TpSolver.csproj", "{A9F017D6-F0FE-4107-A8A0-EE6AF872C06C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VamLib.Tests", "VamLib.Tests\VamLib.Tests.csproj", "{070AFDE8-717E-4A84-8E8E-0D4F30599CD1}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TpSolver.Tests", "TpSolver.Tests\TpSolver.Tests.csproj", "{88ACB388-1935-423D-8AFC-3A8C21C86294}"
 EndProject
 Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B131C715-66E0-4DB4-BB2C-0AB6156CE52E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B131C715-66E0-4DB4-BB2C-0AB6156CE52E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B131C715-66E0-4DB4-BB2C-0AB6156CE52E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B131C715-66E0-4DB4-BB2C-0AB6156CE52E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{070AFDE8-717E-4A84-8E8E-0D4F30599CD1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{070AFDE8-717E-4A84-8E8E-0D4F30599CD1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{070AFDE8-717E-4A84-8E8E-0D4F30599CD1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{070AFDE8-717E-4A84-8E8E-0D4F30599CD1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A9F017D6-F0FE-4107-A8A0-EE6AF872C06C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A9F017D6-F0FE-4107-A8A0-EE6AF872C06C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A9F017D6-F0FE-4107-A8A0-EE6AF872C06C}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{A9F017D6-F0FE-4107-A8A0-EE6AF872C06C}.Debug|x64.Build.0 = Debug|Any CPU
+		{A9F017D6-F0FE-4107-A8A0-EE6AF872C06C}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{A9F017D6-F0FE-4107-A8A0-EE6AF872C06C}.Debug|x86.Build.0 = Debug|Any CPU
+		{A9F017D6-F0FE-4107-A8A0-EE6AF872C06C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A9F017D6-F0FE-4107-A8A0-EE6AF872C06C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A9F017D6-F0FE-4107-A8A0-EE6AF872C06C}.Release|x64.ActiveCfg = Release|Any CPU
+		{A9F017D6-F0FE-4107-A8A0-EE6AF872C06C}.Release|x64.Build.0 = Release|Any CPU
+		{A9F017D6-F0FE-4107-A8A0-EE6AF872C06C}.Release|x86.ActiveCfg = Release|Any CPU
+		{A9F017D6-F0FE-4107-A8A0-EE6AF872C06C}.Release|x86.Build.0 = Release|Any CPU
+		{88ACB388-1935-423D-8AFC-3A8C21C86294}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{88ACB388-1935-423D-8AFC-3A8C21C86294}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{88ACB388-1935-423D-8AFC-3A8C21C86294}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{88ACB388-1935-423D-8AFC-3A8C21C86294}.Debug|x64.Build.0 = Debug|Any CPU
+		{88ACB388-1935-423D-8AFC-3A8C21C86294}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{88ACB388-1935-423D-8AFC-3A8C21C86294}.Debug|x86.Build.0 = Debug|Any CPU
+		{88ACB388-1935-423D-8AFC-3A8C21C86294}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{88ACB388-1935-423D-8AFC-3A8C21C86294}.Release|Any CPU.Build.0 = Release|Any CPU
+		{88ACB388-1935-423D-8AFC-3A8C21C86294}.Release|x64.ActiveCfg = Release|Any CPU
+		{88ACB388-1935-423D-8AFC-3A8C21C86294}.Release|x64.Build.0 = Release|Any CPU
+		{88ACB388-1935-423D-8AFC-3A8C21C86294}.Release|x86.ActiveCfg = Release|Any CPU
+		{88ACB388-1935-423D-8AFC-3A8C21C86294}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
