+++ TpSolver/BfsSearch/Vam.cs
@@ -1,6 +1,6 @@
 ﻿using System.Diagnostics;
 
-namespace VamLib;
+namespace TpSolver.BfsSearch;
 
 public class Vam
 {
